+++ src/plugin/accessories/DoorbellCameraAccessory.ts
@@ -72,9 +72,6 @@ export class DoorbellCameraAccessory extends CameraAccessory {
     if (!this.ring_triggered) {
       this.ring_triggered = true;
       this.platform.log.debug(this.accessory.displayName, 'DoorBell ringing');
-      if (this.cameraConfig.useCachedLocalLivestream && this.streamingDelegate) {
-        this.streamingDelegate.prepareCachedStream();
-      }
       this.doorbellService
         .getCharacteristic(this.platform.Characteristic.ProgrammableSwitchEvent)
         .updateValue(this.platform.Characteristic.ProgrammableSwitchEvent.SINGLE_PRESS);