+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 You can find the complete detailled changelog for every beta release [here](https://github.com/homebridge-eufy-security/plugin/releases).
 
-## 2.2.0-beta.1 (03/08/2022) - latest changes
+## 2.2.0-beta.1 - latest changes
 
 ### Added
 
@@ -10,7 +10,6 @@ You can find the complete detailled changelog for every beta release [here](http
 
 ### Changed
 
-- Adjusting streaming resolution is now by padding the image instead of truncating it
 - If HomeKit Secure Video is enabled, motion detection will not work while a recording is active. This is due to technical limitations for now. (see #6)
 - Removed livestream caching (a.k.a `useCachedLocalLivestream`) since the continued streaming could cause issues with HKSV and motion detection in general